refactor: appendObjectLogTool uses TaskTrellisService
