+++ .trellis/t/open/T-refactor-claimtasktool-to-use.md
@@ -30,6 +30,14 @@ The claimTaskTool currently uses Repository directly. Need to refactor it to use
 
 3. **Migrate and update tests**:
    - **Move business logic tests**: Extract detailed claiming logic tests from `src/tools/__tests__/claimTaskTool.test.ts` and move them to `src/services/local/__tests__/LocalTaskTrellisService.test.ts`
+   - **Create dedicated service function tests**: Create `src/services/local/__tests__/claimTask.test.ts` with comprehensive test coverage for the `claimTask` function, including:
+     - Normal task claiming scenarios
+     - Scope filtering
+     - Force claiming
+     - Error handling (object not found, validation errors, etc.)
+     - Task readiness validation
+     - Priority-based selection
+     - All edge cases and business logic
    - **Update tool tests**: Refactor `src/tools/__tests__/claimTaskTool.test.ts` to:
      - Mock TaskTrellisService instead of Repository
      - Test only that service.claimTask is called with correct parameters
@@ -42,6 +50,7 @@ The claimTaskTool currently uses Repository directly. Need to refactor it to use
 - [ ] Function delegates to service.claimTask with all parameters
 - [ ] server.ts integration updated
 - [ ] Business logic tests moved from tool tests to service tests
+- [ ] Dedicated `src/services/local/__tests__/claimTask.test.ts` created with comprehensive coverage
 - [ ] Tool tests updated to mock service and verify integration only
 - [ ] Service tests cover all claiming logic scenarios
 - [ ] TypeScript compilation passes
