+++ src/tools/appendObjectLogTool.ts
@@ -1,4 +1,5 @@
 import { Repository } from "../repositories";
+import { TaskTrellisService } from "../services/TaskTrellisService";
 
 export const appendObjectLogTool = {
   name: "append_object_log",
@@ -46,6 +47,7 @@ The log creates a permanent audit trail that helps with project retrospectives, 
 } as const;
 
 export async function handleAppendObjectLog(
+  service: TaskTrellisService,
   repository: Repository,
   args: unknown,
 ) {
@@ -54,49 +56,5 @@ export async function handleAppendObjectLog(
     contents: string;
   };
 
-  try {
-    // Load the existing object
-    const existingObject = await repository.getObjectById(id);
-    if (!existingObject) {
-      return {
-        content: [
-          {
-            type: "text",
-            text: `Error: Object with ID '${id}' not found`,
-          },
-        ],
-      };
-    }
-
-    // Create updated object with new log entry appended
-    const updatedObject = {
-      ...existingObject,
-      log: [...existingObject.log, contents],
-    };
-
-    // Save the updated object
-    await repository.saveObject(updatedObject);
-
-    return {
-      content: [
-        {
-          type: "text",
-          text: `Successfully appended to object log: ${JSON.stringify(
-            { id, contents, totalLogEntries: updatedObject.log.length },
-            null,
-            2,
-          )}`,
-        },
-      ],
-    };
-  } catch (error) {
-    return {
-      content: [
-        {
-          type: "text",
-          text: `Error appending to object log: ${error instanceof Error ? error.message : String(error)}`,
-        },
-      ],
-    };
-  }
+  return service.appendObjectLog(repository, id, contents);
 }
